move getVerboseColumns logic into its own file and refactor
